backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
bcrypt==3.2.2
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.2.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stegano==2.0.0
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
opencv-python-headless==4.9.0.80
